+++ packages/app/src/index.d.ts
@@ -1,0 +1,3 @@
+export { default as RainbowFirebaseApp } from './components/App';
+export { default as SideBarNavigation } from './components/SideBarNavigation';
+export { default as SideBarOption } from './components/SideBarOption';