+++ packages/listview/src/components/FloatingSearchButtonIcon/__test__/floatingSearchButtonIcon.spec.js
@@ -1,6 +1,7 @@
 import React from 'react';
 import { mount } from 'enzyme';
-import FloatingSearchButtonIcon from '../index';
+import { ButtonIcon } from 'react-rainbow-components';
+import FloatingSearchButtonIcon from '..';
 import FloatingSearch from '../../FloatingSearch';
 
 jest.mock('../../FloatingSearch', () =>
@@ -10,14 +11,14 @@ jest.mock('../../FloatingSearch', () =>
 describe('<FloatingSearchButtonIcon />', () => {
     it('should render a button and not render FloatingSearch', () => {
         const component = mount(<FloatingSearchButtonIcon />);
-        const button = component.find('ButtonIcon');
+        const button = component.find(ButtonIcon);
         expect(button.exists()).toBe(true);
         expect(component.find(FloatingSearch).prop('isVisible')).toBe(false);
     });
 
     it('should render a FloatingSearch when click on button', () => {
         const component = mount(<FloatingSearchButtonIcon />);
-        const button = component.find('ButtonIcon');
+        const button = component.find(ButtonIcon);
         button.simulate('click');
         expect(component.find(FloatingSearch).prop('isVisible')).toBe(true);
     });