+++ packages/record/docs/stories/RecordDetails.story.js
@@ -1,4 +1,5 @@
 import React, { useState, useEffect } from 'react';
+import { BrowserRouter } from 'react-router-dom';
 import styled from 'styled-components';
 import { Application, Badge } from 'react-rainbow-components';
 import RecordField from '../../src/components/RecordField';
@@ -36,33 +37,40 @@ export const BasicRecordDetails = () => {
     const loading = useChangeLoading();
 
     return (
-        <Application>
-            <Container>
-                <RecordDetails>
-                    <RecordField label="Organization" value="Google" isLoading={loading} />
-                    <RecordField
-                        label="Date & Time"
-                        value={new Date()}
-                        type="dateTime"
-                        isLoading={loading}
-                    />
-                    <RecordField
-                        label="Web page"
-                        value="https://google.com"
-                        href="https://google.com"
-                        type="url"
-                        isLoading={loading}
-                    />
-                    <RecordField label="Price" value={50.5} type="currency" isLoading={loading} />
-                    <RecordField
-                        label="Status"
-                        component={StatusBadge}
-                        isLoading={loading}
-                        value="success"
-                    />
-                </RecordDetails>
-            </Container>
-        </Application>
+        <BrowserRouter>
+            <Application>
+                <Container>
+                    <RecordDetails>
+                        <RecordField label="Organization" value="Google" isLoading={loading} />
+                        <RecordField
+                            label="Date & Time"
+                            value={new Date()}
+                            type="dateTime"
+                            isLoading={loading}
+                        />
+                        <RecordField
+                            label="Web page"
+                            value="https://google.com"
+                            href="https://google.com"
+                            type="url"
+                            isLoading={loading}
+                        />
+                        <RecordField
+                            label="Price"
+                            value={50.5}
+                            type="currency"
+                            isLoading={loading}
+                        />
+                        <RecordField
+                            label="Status"
+                            component={StatusBadge}
+                            isLoading={loading}
+                            value="success"
+                        />
+                    </RecordDetails>
+                </Container>
+            </Application>
+        </BrowserRouter>
     );
 };