+++ packages/record/src/components/RecordField/styled/index.js
@@ -67,17 +67,17 @@ export const ValueContainer = styled.span`
         `
         padding-left: 30px;
     `}
-    ${(props) =>
-        props.type === 'url' &&
-        `
-            > a {
-                color:${props.theme.rainbow.palette.text.main};
+`;
 
-                : hover {
-                    color:${props.theme.rainbow.palette.text.main};
-                }
-            }
-    `};
+export const StyledLink = styled.a`
+    color: ${(props) => props.theme.rainbow.palette.text.main};
+    text-decoration: underline;
+
+    :hover,
+    :active,
+    :visited {
+        color: ${(props) => props.theme.rainbow.palette.text.main};
+    }
 `;
 
 export const StyledLoadingLabel = styled.div`