+++ packages/validation/src/index.d.ts
@@ -1,0 +1,18 @@
+/* eslint-disable @typescript-eslint/no-explicit-any */
+export function isBoolean(value: any): boolean;
+export function isDate(value: any): boolean;
+export function isDateInFuture(value: any): boolean;
+export function isDateInPast(value: any): boolean;
+export function isDefined(value: any): boolean;
+export function isEmail(value: string): boolean;
+export function isEmpty(value: any): boolean;
+export function isEmptyArray(value: any): boolean;
+export function isEmptyObject(value: any): boolean;
+export function isEmptyString(value: any): boolean;
+export function isFunction(value: any): boolean;
+export function isInteger(value: any): boolean;
+export function isNull(value: any): boolean;
+export function isNumber(value: any): boolean;
+export function isPhoneNumber(value: string): boolean;
+export function isPromise(value: any): boolean;
+export function isUndefined(value: any): boolean;